Factor out functions to insert entries into storage
